pyproject.toml
[project]
name = "wqi-research-vit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.10",
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "streamlit>=1.56.0",
    "streamlit-folium>=0.27.1",
    "xgboost>=3.2.0",
]
